NPUW: Move VariableState out of the LoRA header
ov::npuw::VariableState is a general-purpose state over a lifetime-bound
tensor - it deliberately implements no reset() - and util.cpp already uses
it outside any LoRA context. It only happened to live in
llm_lora_states.hpp, which forced the batched element to include a LoRA
header to express "skip the states that must survive the row resets".

Give the class its own variable_state.hpp and leave llm_lora_states.hpp
with the LoRA name patterns. The batched element now depends on the
generic state type only and its reset-filter comment no longer needs to
mention LoRA at all.
